requirements.txt
keyboard
pyautogui
pillow
pygame